extract common project for common stuff
