+++ common/src/main/scala/com/gu/sfl/persistance/PersistenceConfig.scala
@@ -1,0 +1,5 @@
+package com.gu.sfl.persistance
+
+case class PersistenceConfig(app: String, stage: String) {
+  val tableName = s"$app-$stage-articles"
+}
