+++ mobile-save-for-later-user-deletion/src/main/scala/com/gu/sfl/userdeletion/UserDeletionLambda.scala
@@ -1,10 +1,11 @@
 package com.gu.sfl.userdeletion
 
 import com.amazonaws.services.lambda.runtime.events.SQSEvent
-import com.amazonaws.services.sqs.model.Message
-import com.gu.sfl.logging.Logging
-import com.gu.sfl.userdeletion.lib.Jackson.mapper
+import com.gu.sfl.Logging
+import com.gu.sfl.lib.Jackson.mapper
+import com.gu.sfl.persistance.PersistenceConfig
 import com.gu.sfl.userdeletion.model.User
+import coml.gu.sfl.userdeletion.db.SflDynamoDb
 
 import scala.collection.JavaConverters._
 
@@ -13,11 +14,10 @@ object UserDeletionLambda extends Logging {
 
   logger.info("Hello from lambda!")
 
-  //SaveForLaterApp
   val saveForLaterApp = Option(System.getenv("SaveForLaterApp")).getOrElse(sys.error("No main app name configured"))
   val stage = Option(System.getenv("Stage")).getOrElse(sys.error("No main stage configured"))
 
-  val sflDyanamo = new SflDynamoDb(PersistanceConfig(saveForLaterApp, stage))
+  val sflDyanamo = new SflDynamoDb(PersistenceConfig(saveForLaterApp, stage))
 
   def handler(sQSEvent: SQSEvent) {
     val messages = sQSEvent.getRecords.asScala.map(mes => mes).toList
