--- mobile-save-for-later-user-deletion/src/main/scala/com/gu/sfl/userdeletion/logging/Logging.scala
@@ -1,8 +1,0 @@
-package com.gu.sfl.logging
-
-import org.apache.logging.log4j.LogManager
-;
-
-trait Logging {
-  val logger = LogManager.getLogger(this.getClass.getName)
-}
